Movement and mesh loading
Here's the code so far, Josh's classes added.

There's an issue with how it reads the .obj files in (so it won't run
for now), because it's doing it as if it was done on a PC. Josh is aware
of this and is planning to look into it later, once that's fixed I think
everything should work fine, but I'll do some tinkering to make sure.

I put the obj and material files in res/raw, which is where we should
load them from.

Got to head out now.